Boat Database 2012
This is my 2nd year Java 6 degree assignment, it is Object Orientated
Programming and my Objects are to book a Booking, to book a Boat and for
the main menu with a database. I has an age verification with my if,
else if and else loops, it swaps X and Y for the first and second year
of birth, which is to prevent human errors. I wrote two dimensional code
for the temperatures at the different locations. It saves all data into
a text file, so that you can turn off your computer and I wrote an
upload method into the start of the programme so that it automatically
extracts all the data from the text file on your computer and puts it
back into the arrays. I gained 71% distinction at university for this
Java 6 programme and I'm very proud of it.